requirements.txt
python-dotenv==1.0.1
SpeechRecognition==3.14.1
gpiozero==2.0.1
sounddevice==0.5.1
PyAudio==0.2.14
openai==1.60.2
google-search-results==2.4.2
gTTS==2.5.4